# src/openapi.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Blog backend CRUD documentation",
      "description": "This is a basic user API it allows you create, retrieve update and delete users from an in memmory array written in node / express",
      "contact": {
        "email": "n01470339@unf.edu"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000/"
      }
    ],
...
    "paths": {
        "/Posts":{
            "get": {
                "tags":["Posts"],
                "description":"Returns all posts",
                "responses":{
                    "200":{
                        "description": "Returns a list of posts that are currently in the in memmory array",
                        "content":{
                            "application/json": {
                              "schema": {
                                "type": "array",
                                "items": {
                                  "$ref": "#/components/schemas/PostResponse"
                                }
                              }
                            }
                          }
                    }
                }
            },
            "post": {
                "security": [
                {
                    "bearerAuth": []
                }
                ],
                "tags":["Posts"],
                "description":"Creates a new blog post",
                "requestBody":{
                    "content": {
                        "application/json": {
                          "schema": {
                            "$ref": "#/components/schemas/PostRequest"
                          }
                        }
                      },
                      "required": true
                },
                "responses":{
                    "201":{
                        "description": "Post Created",
                        "content":{
                            "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/PostResponse"
                              }
                            }
                          }
                    },
                    "401":{
                        "description":"Unauthorized - Access token is missing or invalid",
                        "content": {
                            "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/ErrorResponse"
                              }
                            }
                          }
                    },
                    "406":{
                        "description":"Not Acceptable: Bad data in the entity IE: Missing Title or Content",
                        "content": {
                            "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/ErrorResponse"
                              }
                            }
                          }
                    }
                }
            }
        },
        "/Posts/{postId}": {
            "get":{
                "tags":["Posts"],
                "description": "Returns a post",
                "parameters":[
                    {
                      "name": "postId",
                      "in": "path",
                      "description": "postId of the post object you are trying to find",
                      "required": true,
                      "schema": {
                        "type": "string"
                      }
                    }
                ],
                "responses": {
                    "200": {
                      "description": "Post",
                      "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostResponse"
                            }
                          }
                      }
                    },
                    "401": {
                        "description": "Post not Found",
                        "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/ErrorResponse"
                              }
                            }
                        }
                      }
                  }
            },
            "patch":{
                "security": [
                {
                    "bearerAuth": []
                }
                ],
                "tags":["Posts"],
                "description": "Updates a specific post",
                "parameters":[
                    {
                      "name": "postId",
                      "in": "path",
                      "description": "postId of the post object you are trying to update",
                      "required": true,
                      "schema": {
                        "type": "string"
                      }
                    }
                ],
                "requestBody": {
                    "content": {
                      "application/json": {
                        "schema": {
                          "$ref": "#/components/schemas/PostRequest"
                        }
                      }
                    },
                    "required": true
                  },
                "responses": {
                    "200": {
                      "description": "Post Updated",
                      "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostResponse"
                            }
                          }
                      }
                    },
                    "401":{
                        "description":"Unauthorized - Access token is missing or invalid",
                        "content": {
                            "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/ErrorResponse"
                              }
                            }
                          }
                    },
                    "404":{
                        "description":"Post not found",
                        "content": {
                            "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/ErrorResponse"
                              }
                            }
                          }
                    }
                }
            },
            "delete":{
                "security": [
                {
                    "bearerAuth": []
                }
                ],
                "tags":["Posts"],
                "description": "Deletes a Post and any related Post Categories",
                "parameters":[
                    {
                      "name": "postId",
                      "in": "path",
                      "description": "postId of the post object you are trying to delete",
                      "required": true,
                      "schema": {
                        "type": "string"
                      }
                    }
                ],
                "responses": {
                    "204": {
                      "description": "Post deleted"
                    },
                    "401":{
                        "description":"Unauthorized - Access token is missing or invalid",
                        "content": {
                            "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/ErrorResponse"
                              }
                            }
                          }
                    },
                    "404":{
                        "description":"Post not found",
                        "content": {
                            "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/ErrorResponse"
                              }
                            }
                          }
                    }
                }
            }
        },
        "/Posts/User/{userId}": {
            "get":{
                "tags":["Posts"],
                "description": "Returns a post array",
                "parameters":[
                    {
                      "name": "userId",
                      "in": "path",
                      "description": "userId on the post object you are trying to find",
                      "required": true,
                      "schema": {
                        "type": "string"
                      }
                    }
                ],
                "responses": {
                    "200": {
                      "description": "Array of available posts for the given user",
                      "content": {
                        "application/json": {
                            "type":"array",
                            "schema": {
                                "$ref": "#/components/schemas/PostResponse"
                            }
                          }
                      }
                    },
                    "404": {
                        "description": "Invalid user / no posts found",
                        "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/ErrorResponse"
                              }
                            }
                        }
                      }
                  }
            }
        },
...
    },
    "components": {
      "securitySchemes":{
        "bearerAuth": {
            "type": "http",
            "scheme": "bearer"
          }
      },
      "schemas": {
...
        "PostResponse": {
          "properties": {
           "postId": {
              "type": "integer"
            },
            "createdDate": {
              "type": "string"
            },
            "title": {
              "type": "string"
            },
            "content": {
              "type": "string"
            },
            "userId": {
              "type": "string"
            },
            "headerImage": {
              "type": "string"
            },
            "lastUpdated": {
              "type": "string"
            }
          }
        },
        "PostRequest": {
          "properties": {
            "title": {
              "type": "string"
            },
            "content": {
              "type": "string"
            },
            "headerImage": {
              "type": "string"
            }
          }
        },
        "ErrorResponse": {
          "properties": {
            "message": {
              "type": "string"
            },
            "status": {
              "type": "string"
            }
          }
        }
      }
    }
  }